Meta tx

+++ contracts/Token.sol
@@ -14,10 +14,12 @@ contract Token is ERC20 {
     constructor (string name, string symbol, uint8 decimals, uint256 totalSupply)
     public
     ERC20 (name, symbol, decimals, totalSupply) {
-        _balances[msg.sender] = totalSupply * (10 ** decimals);
+        _balances[msg.sender] = totalSupply * (uint256(10) ** decimals);
 
-        // TODO: Add vesting schedule
-        // new TokenVesting(beneficiary, start, cliffDuration, duration);
+    // vesting Example
+    // address tokenVesting = new TokenVesting(0x9fB29AAc15b9A4B7F17c3385939b007540f4d791, 1552314037, 10, 2);
+    // _transfer(msg.sender, 0x9fB29AAc15b9A4B7F17c3385939b007540f4d791, 100);
+
     }
 
 }
